add item and code in ventas details

+++ ventas/views.py
@@ -4,7 +4,7 @@ from django.core.exceptions import ValidationError
 from django.db import transaction
 from django.shortcuts import redirect, render
 from django.urls import reverse_lazy
-from django.views.generic import CreateView, DeleteView, DetailView, ListView
+from django.views.generic import CreateView, DeleteView, DetailView, ListView, UpdateView
 
 from productos.models import Product
 from .forms import SaleDetailFormSet, SaleForm
@@ -59,22 +59,27 @@ class SaleCreateView(SalesAccessMixin, CreateView):
 		formset = SaleDetailFormSet(request.POST, instance=self.object)
 
 		if form.is_valid() and formset.is_valid():
-			self.object = form.save()
+			self.object = form.save(commit=False)
+			self.object.seller = request.user
+			self.object.save()
 			formset.instance = self.object
 			formset.save()
 			self.object.calculate_total()
-			from caja.models import CashBox
-			CashBox.validate_day_open(self.object.date)
-			try:
-				self.object.apply_inventory_output()
-			except ValidationError as exc:
-				messages.error(request, str(exc))
-				self.object.delete()
-				return render(request, self.template_name, self.get_context_data(form=form, formset=formset))
 
-			CashBox.register_sale(self.object)
+			if self.object.status == self.object.STATUS_CONFIRMED:
+				from caja.models import CashBox
+				CashBox.validate_day_open(self.object.date)
+				try:
+					self.object.apply_inventory_output()
+				except ValidationError as exc:
+					messages.error(request, str(exc))
+					self.object.delete()
+					return render(request, self.template_name, self.get_context_data(form=form, formset=formset))
+				CashBox.register_sale(self.object)
+				messages.success(request, "Venta registrada exitosamente.")
+			else:
+				messages.success(request, "Proforma guardada exitosamente.")
 
-			messages.success(request, "Venta registrada exitosamente.")
 			return redirect(self.success_url)
 
 		return render(request, self.template_name, self.get_context_data(form=form, formset=formset))
@@ -102,6 +107,78 @@ class SaleCreateView(SalesAccessMixin, CreateView):
 		return super().get_context_data(**kwargs)
 
 
+class SaleUpdateView(SalesAccessMixin, UpdateView):
+	model = Sale
+	form_class = SaleForm
+	template_name = "ventas/sale_form.html"
+	success_url = reverse_lazy("ventas:list")
+
+	def get(self, request, *args, **kwargs):
+		self.object = self.get_object()
+		if self.object.status != Sale.STATUS_PROFORMA:
+			messages.error(request, "Solo se pueden editar ventas en estado proforma.")
+			return redirect(self.success_url)
+		return super().get(request, *args, **kwargs)
+
+	@transaction.atomic
+	def post(self, request, *args, **kwargs):
+		self.object = self.get_object()
+		if self.object.status != Sale.STATUS_PROFORMA:
+			messages.error(request, "Solo se pueden editar ventas en estado proforma.")
+			return redirect(self.success_url)
+
+		form = self.get_form()
+		formset = SaleDetailFormSet(request.POST, instance=self.object)
+
+		if form.is_valid() and formset.is_valid():
+			new_status = form.cleaned_data.get("status")
+			self.object = form.save(commit=False)
+			self.object.save()
+			formset.instance = self.object
+			formset.save()
+			self.object.calculate_total()
+
+			if new_status == Sale.STATUS_CONFIRMED:
+				from caja.models import CashBox
+				CashBox.validate_day_open(self.object.date)
+				try:
+					self.object.apply_inventory_output()
+				except ValidationError as exc:
+					messages.error(request, str(exc))
+					return render(request, self.template_name, self.get_context_data(form=form, formset=formset))
+				CashBox.register_sale(self.object)
+				messages.success(request, "Proforma confirmada como venta exitosamente.")
+			else:
+				messages.success(request, "Proforma actualizada exitosamente.")
+
+			return redirect(self.success_url)
+
+		return render(request, self.template_name, self.get_context_data(form=form, formset=formset))
+
+	def get_context_data(self, **kwargs):
+		if self.request.POST:
+			kwargs["formset"] = SaleDetailFormSet(self.request.POST, instance=self.object)
+		else:
+			kwargs["formset"] = SaleDetailFormSet(instance=self.object)
+		products = Product.objects.select_related("brand", "category").all()
+		kwargs["products_data"] = [
+			{
+				"id": product.id,
+				"name": product.name,
+				"price": float(product.price),
+				"stock": product.stock,
+				"brand": product.brand.name,
+				"category": product.category.name,
+				"size": product.size,
+				"color": product.color,
+				"image": product.image.url if product.image else "",
+			}
+			for product in products
+		]
+		kwargs["is_edit"] = True
+		return super().get_context_data(**kwargs)
+
+
 class SaleDeleteView(SalesAccessMixin, DeleteView):
 	model = Sale
 	template_name = "ventas/sale_confirm_delete.html"
@@ -110,13 +187,16 @@ class SaleDeleteView(SalesAccessMixin, DeleteView):
 	@transaction.atomic
 	def post(self, request, *args, **kwargs):
 		self.object = self.get_object()
-		from caja.models import CashBox
-		try:
-			CashBox.validate_day_open()
-		except ValidationError as exc:
-			messages.error(request, str(exc))
-			return redirect(self.success_url)
-		self.object.restore_inventory_output()
-		CashBox.register_sale_reversal(self.object)
-		messages.warning(request, "Venta eliminada y stock restaurado.")
+		if self.object.status == self.object.STATUS_CONFIRMED:
+			from caja.models import CashBox
+			try:
+				CashBox.validate_day_open()
+			except ValidationError as exc:
+				messages.error(request, str(exc))
+				return redirect(self.success_url)
+			self.object.restore_inventory_output()
+			CashBox.register_sale_reversal(self.object)
+			messages.warning(request, "Venta eliminada y stock restaurado.")
+		else:
+			messages.warning(request, "Proforma eliminada.")
 		return super().post(request, *args, **kwargs)
